+++ test/skyscraper_test.clj
@@ -67,3 +67,22 @@
        "//bar.uk/baz/foo" "https://bar.uk/baz/foo"
        "/baz/foo" "https://foo.pl/baz/foo"
        "foo" "https://foo.pl/bar/foo"))
+
+(defprocessor nil-url-test-processor-root
+  :cache-template "nil-url"
+  :process-fn (fn [res ctx]
+                (for [a (select res [:a])]
+                  {:title (text a), :url (href a), :processor :skyscraper-test/nil-url-test-processor-child})))
+
+(defprocessor nil-url-test-processor-child
+  :cache-template "nil-url/:title"
+  :process-fn (fn [res ctx]
+                [{:info (text (first (select res [:h1])))}]))
+
+(deftest test-nil-url
+  (let [html-main "<html><a href='/sub'>link</a><a name='anchor'>non-link</a></html>"
+        html-child "<html><h1>Info</h1></html>"
+        get (fn [url & args] {:body (if (string/ends-with? url "/sub") html-child html-main)})]
+    (with-redefs [http/get get]
+      (is (= (scrape [{:url "http://localhost/", :processor :skyscraper-test/nil-url-test-processor-root}] :html-cache nil :processed-cache nil)
+             [{:title "link", :info "Info"}])))))